Moved run functions down to ci.objects
