+++ openaddr/ci/__init__.py
@@ -1,8 +1,13 @@
 import logging; _L = logging.getLogger('openaddr.ci')
 
 from ..compat import standard_library
-from .objects import add_job, write_job, read_job, complete_set, update_set_renders
-from .. import jobs, render, __version__
+from .. import jobs, render
+
+from .objects import (
+    add_job, write_job, read_job, complete_set, update_set_renders,
+    add_run, set_run, copy_run, read_completed_set_runs,
+    get_completed_file_run, get_completed_run
+    )
 
 from os.path import relpath, splitext, join, basename
 from datetime import timedelta
@@ -397,11 +402,7 @@ def enqueue_sources(queue, the_set, sources):
 def _update_expected_paths(db, expected_paths, the_set):
     ''' Discard sources from expected_paths set as they appear in runs table.
     '''
-    db.execute('''SELECT source_path FROM runs
-                  WHERE set_id = %s AND status IS NOT NULL''',
-               (the_set.id, ))
-
-    for (source_path, ) in db:
+    for (_, source_path, _, _) in read_completed_set_runs(db, the_set.id):
         _L.debug(u'Discarding {}'.format(source_path))
         expected_paths.discard(source_path)
 
@@ -437,13 +438,9 @@ def render_set_maps(s3, db, the_set):
 def _prepare_render_sources(db, the_set, dirname):
     ''' Dump all non-null set runs into a directory for rendering.
     '''
-    db.execute('''SELECT source_id, source_data, status FROM runs
-                  WHERE set_id = %s AND status IS NOT NULL''',
-               (the_set.id, ))
-    
     good_sources = set()
     
-    for (source_id, source_data, status) in db.fetchall():
+    for (source_id, _, source_data, status) in read_completed_set_runs(db, the_set.id):
         filename = '{source_id}.json'.format(**locals())
         with open(join(dirname, filename), 'w+b') as file:
             content = b64decode(bytes(source_data))
@@ -513,12 +510,7 @@ def is_completed_run(db, run_id, min_datetime):
         # Assume unspecified time zones are UTC.
         min_dtz = min_datetime.replace(tzinfo=tzutc())
 
-    db.execute('''SELECT id, status FROM runs
-                  WHERE id = %s AND status IS NOT NULL
-                    AND datetime_tz >= %s''',
-               (run_id, min_dtz))
-    
-    completed_run = db.fetchone()
+    completed_run = get_completed_run(db, run_id, min_dtz)
     
     if completed_run:
         _L.debug('Found completed run {0} ({1}) since {min_datetime}'.format(*completed_run, **locals()))
@@ -526,73 +518,6 @@ def is_completed_run(db, run_id, min_datetime):
         _L.debug('No completed run {run_id} since {min_datetime}'.format(**locals()))
     
     return bool(completed_run is not None)
-
-def add_run(db):
-    ''' Reserve a row in the runs table and return its new ID.
-    '''
-    db.execute("INSERT INTO runs (datetime_tz) VALUES (NOW())")
-    db.execute("SELECT currval('ints')")
-    
-    (run_id, ) = db.fetchone()
-    
-    return run_id
-
-def set_run(db, run_id, filename, file_id, content_b64, run_state, run_status,
-            job_id, worker_id, commit_sha, set_id):
-    ''' Populate an identitified row in the runs table.
-    '''
-    db.execute('''UPDATE runs SET
-                  source_path = %s, source_data = %s, source_id = %s,
-                  state = %s::json, status = %s, worker_id = %s,
-                  code_version = %s, job_id = %s, commit_sha = %s,
-                  set_id = %s, datetime_tz = NOW()
-                  WHERE id = %s''',
-               (filename, content_b64, file_id,
-               json.dumps(run_state), run_status, worker_id,
-               __version__, job_id, commit_sha,
-               set_id, run_id))
-
-def copy_run(db, run_id, job_id, commit_sha, set_id):
-    ''' Duplicate a previous run and return its new ID.
-    
-        Use new values for job ID, commit SHA, and set ID.
-    '''
-    db.execute('''INSERT INTO runs
-                  (copy_of, source_path, source_id, source_data, state, status,
-                   worker_id, code_version, job_id, commit_sha, set_id, datetime_tz)
-                  SELECT id, source_path, source_id, source_data, state, status,
-                         worker_id, code_version, %s, %s, %s, NOW()
-                  FROM runs
-                  WHERE id = %s''',
-               (job_id, commit_sha, set_id, run_id))
-
-    db.execute("SELECT currval('ints')")
-    
-    (run_id, ) = db.fetchone()
-    
-    return run_id
-
-def get_previously_completed_run(db, file_id):
-    ''' Look for an existing run on this file ID within the reuse timeout limit.
-    '''
-    interval = '{} seconds'.format(RUN_REUSE_TIMEOUT.seconds + RUN_REUSE_TIMEOUT.days * 86400)
-    
-    db.execute('''SELECT id, state, status FROM runs
-                  WHERE source_id = %s
-                    AND datetime_tz > NOW() - INTERVAL %s
-                    AND status IS NOT NULL
-                    AND copy_of IS NULL
-                  ORDER BY id DESC LIMIT 1''',
-               (file_id, interval))
-    
-    previous_run = db.fetchone()
-    
-    if previous_run:
-        _L.debug('Found previous run {0} ({2}) for file {file_id}'.format(*previous_run, **locals()))
-    else:
-        _L.debug('No previous run for file {file_id}'.format(**locals()))
-
-    return previous_run
 
 def update_job_status(db, job_id, job_url, filename, run_status, results, github_auth):
     '''
@@ -650,7 +575,8 @@ def pop_task_from_taskqueue(s3, task_queue, done_queue, due_queue, output_dir):
         passed_on_kwargs = {k: task.data.get(k) for k in passed_on_keys}
         passed_on_kwargs['worker_id'] = hex(getnode()).rstrip('L')
 
-        previous_run = get_previously_completed_run(db, task.data.get('file_id'))
+        interval = '{} seconds'.format(RUN_REUSE_TIMEOUT.seconds + RUN_REUSE_TIMEOUT.days * 86400)
+        previous_run = get_completed_file_run(db, task.data.get('file_id'), interval)
     
         if previous_run:
             # Make a copy of the previous run.
