+++ openaddr/ci/objects.py
@@ -1,5 +1,6 @@
 import logging; _L = logging.getLogger('openaddr.ci.objects')
 
+from .. import __version__
 import json
 
 class Job:
@@ -147,3 +148,87 @@ def read_sets(db, past_id):
                (past_id, ))
     
     return [Set(*row) for row in db.fetchall()]
+
+def add_run(db):
+    ''' Reserve a row in the runs table and return its new ID.
+    '''
+    db.execute("INSERT INTO runs (datetime_tz) VALUES (NOW())")
+    db.execute("SELECT currval('ints')")
+    
+    (run_id, ) = db.fetchone()
+    
+    return run_id
+
+def set_run(db, run_id, filename, file_id, content_b64, run_state, run_status,
+            job_id, worker_id, commit_sha, set_id):
+    ''' Populate an identitified row in the runs table.
+    '''
+    db.execute('''UPDATE runs SET
+                  source_path = %s, source_data = %s, source_id = %s,
+                  state = %s::json, status = %s, worker_id = %s,
+                  code_version = %s, job_id = %s, commit_sha = %s,
+                  set_id = %s, datetime_tz = NOW()
+                  WHERE id = %s''',
+               (filename, content_b64, file_id,
+               json.dumps(run_state), run_status, worker_id,
+               __version__, job_id, commit_sha,
+               set_id, run_id))
+
+def copy_run(db, run_id, job_id, commit_sha, set_id):
+    ''' Duplicate a previous run and return its new ID.
+    
+        Use new values for job ID, commit SHA, and set ID.
+    '''
+    db.execute('''INSERT INTO runs
+                  (copy_of, source_path, source_id, source_data, state, status,
+                   worker_id, code_version, job_id, commit_sha, set_id, datetime_tz)
+                  SELECT id, source_path, source_id, source_data, state, status,
+                         worker_id, code_version, %s, %s, %s, NOW()
+                  FROM runs
+                  WHERE id = %s''',
+               (job_id, commit_sha, set_id, run_id))
+
+    db.execute("SELECT currval('ints')")
+    
+    (run_id, ) = db.fetchone()
+    
+    return run_id
+
+def get_completed_file_run(db, file_id, interval):
+    ''' Look for an existing run on this file ID within the reuse timeout limit.
+    '''
+    db.execute('''SELECT id, state, status FROM runs
+                  WHERE source_id = %s
+                    AND datetime_tz > NOW() - INTERVAL %s
+                    AND status IS NOT NULL
+                    AND copy_of IS NULL
+                  ORDER BY id DESC LIMIT 1''',
+               (file_id, interval))
+    
+    previous_run = db.fetchone()
+    
+    if previous_run:
+        _L.debug('Found previous run {0} ({2}) for file {file_id}'.format(*previous_run, **locals()))
+    else:
+        _L.debug('No previous run for file {file_id}'.format(**locals()))
+
+    return previous_run
+
+def get_completed_run(db, run_id, min_dtz):
+    '''
+    '''
+    db.execute('''SELECT id, status FROM runs
+                  WHERE id = %s AND status IS NOT NULL
+                    AND datetime_tz >= %s''',
+               (run_id, min_dtz))
+    
+    return db.fetchone()
+
+def read_completed_set_runs(db, set_id):
+    '''
+    '''
+    db.execute('''SELECT source_id, source_path, source_data, status FROM runs
+                  WHERE set_id = %s AND status IS NOT NULL''',
+               (set_id, ))
+    
+    return list(db.fetchall())
